# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Demo Loans System",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "/api/{version}",
            "description": "Api for loans system",
            "variables": {
                "version": {
                    "enum": [
                        "v1"
                    ],
                    "default": "v1"
                }
            }
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user to get token",
                "description": "`user@example.com / 123456`",
                "operationId": "99d66635c4992aeaa6aa44ff653d0563",
                "parameters": [
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "description": "Request Header",
                        "schema": {
                            "type": "string",
                            "default": "XMLHttpRequest"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "402": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Create new user",
                "description": "Create new user account for login",
                "operationId": "67fd37db7cecd428148771f4e12fb15a",
                "parameters": [
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "description": "Request Header",
                        "schema": {
                            "type": "string",
                            "default": "XMLHttpRequest"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/users"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/users"
                                }
                            }
                        }
                    },
                    "402": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "description": "Logout user",
                "operationId": "6abf3b64a4bc7838d56346f05a5153af",
                "parameters": [
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "description": "Request Header",
                        "schema": {
                            "type": "string",
                            "default": "XMLHttpRequest"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "402": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "users": {
                "title": "User Model",
                "properties": {
                    "first_name": {
                        "schema": "User",
                        "description": "First Name",
                        "type": "string",
                        "example": "UserRepo"
                    },
                    "last_name": {
                        "schema": "User",
                        "description": "Last Name",
                        "type": "string",
                        "example": "A"
                    },
                    "email": {
                        "schema": "User",
                        "description": "Email address",
                        "type": "string",
                        "example": "user1@email.com"
                    },
                    "password": {
                        "schema": "User",
                        "description": "password",
                        "type": "string",
                        "example": "123321"
                    },
                    "personal_code": {
                        "schema": "User",
                        "description": "0123456789",
                        "type": "string",
                        "example": "0123456789"
                    },
                    "phone": {
                        "schema": "User",
                        "description": "0912345678",
                        "type": "string",
                        "example": "0912345678"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "description": "Header value: `Bearer {access_token}`",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        []
    ]
}